// src/ocm/ocmv2.d.ts
export type ComponentName = string;
export type RelaxedSemver = string;
export type OciRepositoryContext = RepositoryContext & {
  baseUrl: string;
  componentNameMapping?: "urlPath" | "sha256-digest";
  type?: "ociRegistry" | "OCIRegistry";
  [k: string]: unknown;
};

/**
 * Open Component Model v2 schema
 */
export interface HttpsGardenerCloudSchemasComponentDescriptorV2 {
  meta: Meta;
  component: Component;
  signatures?: Signature[];
  [k: string]: unknown;
}
/**
 * component descriptor metadata
 */
export interface Meta {
  schemaVersion: string;
  [k: string]: unknown;
}
/**
 * a component
 */
export interface Component {
  name: ComponentName;
  version: RelaxedSemver;
  creationTime?: string;
  repositoryContexts: OciRepositoryContext[];
  provider: string;
  labels?: Label[];
  sources: SourceDefinition[];
  componentReferences: ComponentReference[];
  resources: (ResourceType | OciImageResource | GenericResource)[];
  [k: string]: unknown;
}
export interface RepositoryContext {
  type: string;
  [k: string]: unknown;
}
export interface Label {
  name: string;
  value: unknown;
  version?: {
    [k: string]: unknown;
  };
  signing?: boolean;
  [k: string]: unknown;
}
export interface SourceDefinition {
  name: string;
  extraIdentity?: IdentityAttribute;
  version: RelaxedSemver;
  type: string;
  labels?: Label[];
  access: Access | GithubAccess | HttpAccess;
  [k: string]: unknown;
}
export interface IdentityAttribute {
  [k: string]: unknown;
}
/**
 * base type for accesses (for extensions)
 */
export interface Access {
  [k: string]: unknown;
}
export interface GithubAccess {
  type: "github";
  repoUrl: string;
  ref: string;
  commit?: string;
  [k: string]: unknown;
}
export interface HttpAccess {
  type: "http";
  url: string;
  [k: string]: unknown;
}
/**
 * a reference to a component
 */
export interface ComponentReference {
  componentName: ComponentName;
  name: string;
  extraIdentity?: IdentityAttribute;
  version: RelaxedSemver;
  labels?: Label[];
  digest?: null | DigestSpec;
  [k: string]: unknown;
}
export interface DigestSpec {
  hashAlgorithm: string;
  normalisationAlgorithm: string;
  value: string;
  [k: string]: unknown;
}
/**
 * base type for resources
 */
export interface ResourceType {
  name: string;
  extraIdentity?: IdentityAttribute;
  version: RelaxedSemver;
  type: string;
  srcRefs?: SrcRef[];
  relation: "local" | "external";
  labels?: Label[];
  access: Access | OciBlobAccess | LocalFilesystemBlobAccess | LocalOciBlobAccess;
  digest?: null | DigestSpec;
  [k: string]: unknown;
}
/**
 * a reference to a (component-local) source
 */
export interface SrcRef {
  identitySelector?: IdentityAttribute;
  labels?: Label[];
  [k: string]: unknown;
}
export interface OciBlobAccess {
  type: "ociBlob";
  /**
   * A oci reference to the manifest
   */
  ref?: string;
  /**
   * The media type of the object this access refers to
   */
  mediaType?: string;
  /**
   * The digest of the targeted content
   */
  digest?: string;
  /**
   * The size in bytes of the blob
   */
  size?: number;
  [k: string]: unknown;
}
export interface LocalFilesystemBlobAccess {
  type: "localFilesystemBlob";
  /**
   * filename of the blob that is located in the "blobs" directory
   */
  filename: string;
  [k: string]: unknown;
}
export interface LocalOciBlobAccess {
  type: "localOciBlob";
  /**
   * digest of the layer within the current component descriptor
   */
  digest?: string;
  [k: string]: unknown;
}
export interface OciImageResource {
  name: string;
  extraIdentity?: IdentityAttribute;
  version: RelaxedSemver;
  type: "ociImage";
  labels?: Label[];
  access: OciImageAccess;
  digest?: null | DigestSpec;
  [k: string]: unknown;
}
export interface OciImageAccess {
  type: "ociRegistry";
  imageReference: string;
  [k: string]: unknown;
}
export interface GenericResource {
  name: string;
  extraIdentity?: IdentityAttribute;
  version: RelaxedSemver;
  type: "generic";
  labels?: Label[];
  access: GenericAccess;
  digest?: null | DigestSpec;
  [k: string]: unknown;
}
export interface GenericAccess {
  type: "generic";
  [k: string]: unknown;
}
export interface Signature {
  name: string;
  digest: DigestSpec;
  signature: SignatureSpec;
  [k: string]: unknown;
}
export interface SignatureSpec {
  algorithm: string;
  value: string;
  /**
   * The media type of the signature value
   */
  mediaType: string;
  [k: string]: unknown;
}